backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.108.0
h11==0.14.0
httptools==0.6.1
idna==3.6
mariadb==1.1.9
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
